Improve Provenance tests
