+++ test/onnx/onnx_import_provenance.in.cpp
@@ -22,6 +22,7 @@
 #include "ngraph/pass/manager.hpp"
 #include "ngraph/pass/opset0_downgrade.hpp"
 #include "ngraph/provenance.hpp"
+#include "util/provenance_enabler.hpp"
 #include "util/test_control.hpp"
 #include "util/type_prop.hpp"
 
@@ -68,6 +69,8 @@ void test_provenance_tags(const std::shared_ptr<Function> function,
 
 NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_only_output)
 {
+    test::ProvenanceEnabler provenance_enabler;
+
     // the Add node in the model does not have a name,
     // only its output name should be found in the provenance tags
     const auto function = onnx_import::import_onnx_model(
@@ -77,6 +80,8 @@ NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_only_output)
 
 NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_node_name_and_outputs)
 {
+    test::ProvenanceEnabler provenance_enabler;
+
     const auto function = onnx_import::import_onnx_model(
         file_util::path_join(SERIALIZED_ZOO, "onnx/provenance_node_name_and_outputs.prototxt"));
     test_provenance_tags<default_opset::Add>(function, "<ONNX Add (Add_node -> output_of_add)>");
@@ -84,6 +89,8 @@ NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_node_name_and_outputs)
 
 NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_multiple_outputs_op)
 {
+    test::ProvenanceEnabler provenance_enabler;
+
     const auto function = onnx_import::import_onnx_model(
         file_util::path_join(SERIALIZED_ZOO, "onnx/provenance_multiple_outputs_op.prototxt"));
     test_provenance_tags<default_opset::TopK>(function, "<ONNX TopK (TOPK -> values, indices)>");
@@ -91,6 +98,8 @@ NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_multiple_outputs_op)
 
 NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_tagging_constants)
 {
+    test::ProvenanceEnabler provenance_enabler;
+
     const auto function = onnx_import::import_onnx_model(
         file_util::path_join(SERIALIZED_ZOO, "onnx/provenance_input_tags.prototxt"));
     test_provenance_tags<default_opset::Constant>(function,
@@ -99,6 +108,8 @@ NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_tagging_constants)
 
 NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_tagging_parameters)
 {
+    test::ProvenanceEnabler provenance_enabler;
+
     const auto function = onnx_import::import_onnx_model(
         file_util::path_join(SERIALIZED_ZOO, "onnx/provenance_input_tags.prototxt"));
     test_provenance_tags<default_opset::Parameter>(function, "<ONNX Input (input_B) Shape:{}>");
@@ -106,7 +117,7 @@ NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_tagging_parameters)
 
 NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_tag_downgrade_pass)
 {
-    set_provenance_enabled(true);
+    test::ProvenanceEnabler provenance_enabler;
 
     const auto function = onnx_import::import_onnx_model(
         file_util::path_join(SERIALIZED_ZOO, "onnx/provenance_downgrade_topk.prototxt"));
