+++ test/provenance.cpp
@@ -30,6 +30,7 @@
 #include "ngraph/pass/opset0_downgrade.hpp"
 #include "ngraph/pass/opset1_upgrade.hpp"
 #include "ngraph/provenance.hpp"
+#include "util/provenance_enabler.hpp"
 
 using namespace std;
 using namespace ngraph;
@@ -39,18 +40,7 @@ using ProvSet = std::unordered_set<std::string>;
 
 TEST(provenance, provenance)
 {
-    class ProvenanceEnabler
-    {
-    public:
-        ProvenanceEnabler()
-        {
-            saved_enable_state = get_provenance_enabled();
-            set_provenance_enabled(true);
-        }
-        ~ProvenanceEnabler() { set_provenance_enabled(saved_enable_state); }
-    private:
-        bool saved_enable_state;
-    } provenance_enabler;
+    test::ProvenanceEnabler provenance_enabler;
 
     //
     // Before:
@@ -406,7 +396,7 @@ TEST(provenance, builder)
 
 TEST(provenance, fused_copy_origin_tags)
 {
-    set_provenance_enabled(true);
+    test::ProvenanceEnabler provenance_enabler;
 
     auto p1 = make_shared<op::Parameter>(element::f32, PartialShape{2, 3, 4});
     p1->add_provenance_tag("P1");
@@ -439,7 +429,7 @@ TEST(provenance, fused_copy_origin_tags)
 
 TEST(provenance, fused_decomposition_tag)
 {
-    set_provenance_enabled(true);
+    test::ProvenanceEnabler provenance_enabler;
 
     auto p1 = make_shared<op::Parameter>(element::f32, PartialShape{2, 3, 4});
     auto fused_op = make_shared<op::MVN>(p1);
@@ -556,7 +546,7 @@ TEST(provenance, scaled_quantize_concat_unsigned)
 
 TEST(provenance, opset1_upgrade_pass_topk)
 {
-    set_provenance_enabled(true);
+    test::ProvenanceEnabler provenance_enabler;
 
     const size_t axis = 2;
     const size_t k = 10;
@@ -585,7 +575,7 @@ TEST(provenance, opset1_upgrade_pass_topk)
 
 TEST(provenance, opset0_downgrade_pass_topk)
 {
-    set_provenance_enabled(true);
+    test::ProvenanceEnabler provenance_enabler;
 
     const auto data = make_shared<op::Parameter>(element::i32, Shape{5, 10, 15});
     const int32_t k = 10;
@@ -618,7 +608,7 @@ TEST(provenance, opset0_downgrade_pass_topk)
 
 TEST(provenance, opset1_upgrade_pass_graph)
 {
-    set_provenance_enabled(true);
+    test::ProvenanceEnabler provenance_enabler;
 
     auto x = make_shared<op::Parameter>(element::i32, PartialShape{2, 3, 4});
     auto y = make_shared<op::Parameter>(element::i32, PartialShape{2, 3, 4});
@@ -661,7 +651,7 @@ TEST(provenance, opset1_upgrade_pass_graph)
 
 TEST(provenance, opset0_downgrade_pass_graph)
 {
-    set_provenance_enabled(true);
+    test::ProvenanceEnabler provenance_enabler;
 
     auto x = make_shared<op::Parameter>(element::i32, PartialShape{2, 3, 4});
     auto y = make_shared<op::Parameter>(element::i32, PartialShape{2, 3, 4});
